cloud: Move aggregation in a dedicated file
